fix(schema-importer): fix warning due to Netty's native transport library
When we run a container of schema-importer image of M1 Mac, we got a
warnings as follows:

[main] WARN com.datastax.driver.core.NettyUtil - Found Netty's native epoll transport in the classpath, but epoll is not available. Using NIO instead.
java.lang.UnsatisfiedLinkError: could not load a native library: netty_transport_native_epoll_aarch_64
	at io.netty.util.internal.NativeLibraryLoader.load(NativeLibraryLoader.java:239)

This is due to the missing of Netty's epoll library for Arm architecture.
To solve this, this commits add an dependencies on amd64 and arm64
explicitly.